backend/app/assignment_assistant/requirements.txt
openai>=1.6.0
python-dotenv
pdf2image
pytesseract
Pillow